Substitute dependency in example project
