+++ examples/example-project/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("org.jetbrains.kotlin.jvm") version "1.8.21" apply false
+}
