+++ examples/example-project/settings.gradle.kts
@@ -1,9 +1,6 @@
-plugins {
-    id("org.gradle.toolchains.foojay-resolver-convention") version "0.4.0"
-}
-
 rootProject.name = "example-project"
-include("app")
+
+include(":app")
 
 dependencyResolutionManagement {
     repositories {
