+++ gradle.properties
@@ -3,3 +3,5 @@ org.gradle.caching=true
 version=SNAPSHOT
 # Must be later than 2022.1
 gradle.enterprise.version=2022.4
+group=com.github.gabrielfeo
+artifact=gradle-enterprise-api-kotlin
