+++ library/.openapi-generator-ignore
