+++ library/src/main/kotlin/com/gabrielfeo/gradle/enterprise/api/internal/operator/withGradleAttributes.kt
