+++ library/src/test/kotlin/com/gabrielfeo/gradle/enterprise/api/internal/operator/WithGradleAttributesTest.kt
