+++ library/src/test/resources/gradle-attributes-response.json
